Measure how much is tagged
This is the initial commit to track the number of tagged & untagged
items in statsd. This will allow us to put a graph on the dashboard.

The script is intended to be run regularly. We'll probably create a
jenkins job triggered by an hourly cron job.

I've used Typhoeus because we're planning to make more requests and it
could be handy to have something that executes in parallel.

https://trello.com/c/NEYdRp4c