Include content only covered by single sitewide taxonomy
In the previous search request, we surface any content tagged to _any_
taxon. Potentially, these taxons might not yet be tagged to the Topic
Taxonomy and would incorrectly skew the content coverage metric.

Now the search is filtered to root taxons and their ancestors to
accurately count the content that is tagged to the Topic Taxonomy by
using the part_of_taxonomy_tree field.